addition of custom emotes

+++ src/Preconditions/Parameter/Cooldown.cs
@@ -10,7 +10,7 @@ namespace FFA.Preconditions.Parameter
     {
         public TimeSpan CooldownLength { get; }
 
-        public CooldownAttribute(int hours)
+        public CooldownAttribute(double hours)
         {
             CooldownLength = TimeSpan.FromHours(hours);
         }
